removing unused functions, separating legacy but still used function
